requirements.txt
3to2==1.1.1
absl-py==0.9.0
aiofiles==0.4.0
aiohttp==3.6.2
alabaster==0.7.11
alembic==1.3.0
altgraph==0.16.1
amqp==2.6.0
aniso8601==8.0.0
ann-visualizer==2.5
apipkg==1.5
appdirs==1.4.3
applib==1.2
APScheduler==3.6.3
argh==0.26.2
args==0.1.0
astor==0.6.2
astroid==1.6.2
async-timeout==3.0.1
atomicwrites==1.2.1
attrs==18.2.0
audioread==2.1.5
autopep8==1.4
Babel==2.6.0
bcrypt==3.1.6
beautifulsoup4==4.8.2
billiard==3.6.3.0
bleach==1.5.0
blinker==1.4
boto==2.49.0
boto3==1.9.4
botocore==1.12.4
branca==0.2.0
bs4==0.0.1
buildozer==0.39
bz2file==0.98
cachetools==4.0.0
CausalInference==0.1.2
celery==4.4.6
certifi==2018.4.16
cffi==1.14.4
chardet==3.0.4
click==6.7
clint==0.5.1
cloudpickle==1.3.0
cmake==3.15.3
cmdln==2.0.0
CodeIntel==2.0.0
colorama==0.3.9
Command==0.1.0
configobj==5.0.6
configparser==3.5.0
contextlib2==0.5.5
cryptography==2.8
cycler==0.10.0
cymem==1.31.2
Cython==0.28.5
cytoolz==0.9.0.1
decorator==4.2.1
defusedxml==0.6.0
dill==0.2.8.2
Django==1.11
dlib==19.7.0
dnspython==1.16.0
docutils==0.14
en-core-web-sm==2.0.0
entrypoints==0.2.3
esprima==4.0.1
eventlet==0.26.1
execnet==1.5.0
face-recognition==1.2.3
face-recognition-models==0.3.0
fitz==0.0.1.dev2
Flask==1.1.1
Flask-Admin==1.5.4
Flask-APScheduler==1.11.0
Flask-Bcrypt==0.7.1
flask-buzz==0.1.14
Flask-Cors==3.0.7
Flask-JWT-Extended==3.17.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-marshmallow==0.10.1
Flask-Material==0.1.1
Flask-Migrate==2.5.2
Flask-MySQL==1.4.0
Flask-MySQLdb==0.2.0
flask-praetorian==1.0.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.3.2
flask-swaggerui==0.0.1
Flask-User==1.0.2.0
Flask-WTF==0.14.2
floyd-cli==0.10.39
folium==0.5.0
funcsigs==1.0.2
future==0.18.2
gast==0.2.2
gensim==3.5.0
geocoder==1.38.1
geographiclib==1.50
geopy==1.21.0
gitdb2==2.0.5
GitPython==2.1.11
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
gql==0.1.0
graphql-core==2.1
graphviz==0.8.4
greenlet==0.4.16
grpcio==1.27.2
gTTS==2.1.1
gTTS-token==1.1.3
gunicorn==20.0.4
gym==0.17.0
h5py==2.8.0
helper==2.4.2
html-parser==0.2
html5lib==0.9999999
httplib2==0.11.3
idna==2.6
idna-ssl==1.1.0
imageio==2.4.1
imagesize==1.0.0
importlib-metadata==1.7.0
Inflector==2.0.12
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
isodate==0.6.0
isort==4.3.4
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10.3
jmespath==0.9.3
joblib==0.11
jsonschema==2.6.0
keyring==13.2.1
Kivy==1.11.1
Kivy-Garden==0.1.4
kiwisolver==1.0.1
kombu==4.6.11
lazy-object-proxy==1.3.1
libclang-py3==3.9.0
librosa==0.6.0
llvmlite==0.22.0
loader==2017.9.11
Logbook==1.5.3
logcc==2018.3.1
lxml==4.2.1
macholib==1.11
Mako==1.1.0
Markdown==2.6.11
MarkupSafe==1.0
marshmallow==2.15.1
marshmallow-sqlalchemy==0.19.0
matplotlib==2.2.0
mccabe==0.6.1
mistune==0.8.3
mitdeeplearning==0.1.2
mlp==1.1
mock==2.0.0
monotonic==1.5
more-itertools==4.3.0
moviepy==0.2.3.5
msgpack==0.5.6
msgpack-numpy==0.4.3.1
multidict==4.7.5
murmurhash==0.28.0
mysqlclient==1.4.1
Naked==0.1.31
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nibabel==2.3.0
nipype==1.1.2
nltk==3.2.5
notebook==5.4.0
numba==0.37.0
numexpr==2.6.4
numpy==1.16.4
numpydoc==0.8.0
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
onnx==1.6.0
opencv-contrib-python==4.0.0.21
opencv-python==3.4.4.19
opt-einsum==3.1.0
ordereddict==1.1
packaging==17.1
paho-mqtt==1.5.0
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
passlib==1.7.1
pathlib==1.0.1
pathlib2==2.3.2
pathtools==0.1.2
patsy==0.5.0
pbr==4.2.0
pdfkit==0.6.1
pefile==2019.4.18
pendulum==2.0.5
pexpect==4.7.0
pickleshare==0.7.4
Pillow==6.1.0
pipenv==11.7.1
plac==0.9.6
pluggy==0.7.1
ply==3.11
plyer==1.4.2
pony==0.7.12
preshed==1.0.1
promise==2.2.1
prompt-toolkit==1.0.14
protobuf==3.11.3
prov==1.5.0
psutil==5.4.6
psycopg2==2.7.5
ptyprocess==0.6.0
py==1.6.0
py-buzz==0.3.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.4.0
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.9.7
pycryptodomex==3.9.2
pydot==1.2.4
pydotplus==2.0.2
pyflakes==2.0.0
pygame==1.9.6
pyglet==1.5.0
Pygments==2.2.0
PyInquirer==1.0.3
PyInstaller==3.4
PyJWT==1.7.1
pylib==1.0.11
pylint==1.8.3
PyMySQL==0.9.3
pyparsing==2.2.0
PyPDF2==1.26.0
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.9.2
pyreadline==2.1
PySimpleGUI==3.27.0
pysimplegui-exemaker==1.2.0
pytest==3.8.0
pytest-forked==0.2
pytest-xdist==1.23.0
python-dateutil==2.7.0
python-dotenv==0.10.3
python-editor==1.0.4
python-engineio==3.13.1
python-socketio==4.6.0
pytz==2018.4
pytzdata==2019.3
PyWavelets==0.5.2
pywin32-ctypes==0.1.2
pywinpty==0.5.1
pyxnat==0.0.0
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.2
ratelim==0.1.6
raven==6.7.0
rdflib==4.2.2
redis==3.5.3
regex==2017.4.5
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
resampy==0.2.0
rope==0.10.7
rq==1.4.3
rsa==4.0
Rx==1.6.1
s3transfer==0.1.13
scikit-image==0.13.1
scikit-learn==0.19.1
scikit-optimize==0.5.2
scipy==1.4.1
seaborn==0.8.1
selenium==3.141.0
Send2Trash==1.5.0
sentry-sdk==0.5.5
sh==1.12.14
shellescape==3.8.1
shortuuid==0.5.0
simplegeneric==0.8.1
SimpleITK==1.2.0
simplejson==3.16.0
sip==4.19.8
six==1.14.0
sklearn==0.0
smart-open==1.6.0
smmap2==2.0.5
snowballstemmer==1.2.1
soupsieve==2.0
spacy==2.0.12
Sphinx==1.7.6
sphinxcontrib-websupport==1.1.0
spyder==3.3.0
spyder-kernels==0.2.4
SQLAlchemy==1.3.1
statsmodels==0.9.0
sublime==0.1.0
tables==3.4.2
tabulate==0.8.2
tb-nightly==1.14.0a20190301
telegram==0.0.1
termcc==2019.8.2
termcolor==1.1.0
terminado==0.8.1
terminaltables==3.1.0
tesseract==0.1.3
testpath==0.3.1
thinc==6.10.3
toolz==0.9.0
torch==1.0.0
torchvision==0.2.1
tornado==5.0
tqdm==4.25.0
traitlets==4.3.2
traits==4.6.0
typing-extensions==3.7.4.1
tzlocal==2.1
ujson==1.35
urllib3==1.22
vine==1.3.0
virtualenv==15.1.0
virtualenv-clone==0.3.0
wandb==0.6.29
watchdog==0.9.0
wcwidth==0.1.7
websockets==8.1
Werkzeug==0.16.0
widgetsnbextension==3.1.4
wrap==2018.3.1
wrapt==1.10.11
WTForms==2.2.1
xlrd==1.1.0
yarl==1.4.2
zipp==3.1.0
zope.cachedescriptors==4.3.1
zope.event==4.5.0
zope.interface==5.2.0
